BACKEND/auth_system/requirements.txt
absl-py==2.3.1
accelerate==0.30.1
addict==2.4.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
albucore==0.0.24
albumentations==2.0.8
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
asgiref==3.9.1
attrs==25.3.0
av==15.0.0
black==25.1.0
blinker==1.9.0
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
cloths-segmentation==0.0.2
cloudpickle==3.1.1
colorama==0.4.6
contourpy==1.3.3
cycler==0.12.1
detectron2==0.6
diffusers==0.27.2
Django==5.2.4
django-cors-headers==4.7.0
django-environ==0.12.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.1
dnspython==2.7.0
efficientnet_pytorch==0.7.1
filelock==3.18.0
Flask==3.1.1
flask-ngrok==0.0.25
fonttools==4.59.0
frozenlist==1.7.0
fsspec==2025.7.0
fvcore==0.1.5.post20221221
gitdb==4.0.12
GitPython==3.1.45
grpcio==1.74.0
huggingface-hub==0.23.1
hydra-core==1.3.2
idna==3.10
iglovikov-helper-functions==0.0.53
imagecorruptions==1.1.2
imageio==2.37.0
importlib_metadata==8.7.0
iopath==0.1.9
itsdangerous==2.2.0
Jinja2==3.1.3
joblib==1.5.1
jpeg4py==0.1.4
kiwisolver==1.4.8
lazy_loader==0.4
lightning-utilities==0.15.0
Markdown==3.8.2
MarkupSafe==2.1.5
matplotlib==3.10.5
mpmath==1.3.0
multidict==6.6.3
munch==4.0.0
mypy_extensions==1.1.0
networkx==3.3
ninja==1.11.1.4
numpy==1.24.4
omegaconf==2.3.0
opencv-python==4.8.1.78
opencv-python-headless==4.8.1.78
packaging==25.0
pandas==2.3.1
pathspec==0.12.1
peft==0.7.1
pillow==11.3.0
platformdirs==4.3.8
portalocker==3.2.0
pretrainedmodels==0.7.4
propcache==0.3.2
protobuf==6.31.1
psutil==7.0.0
psycopg2-binary==2.9.10
pycocotools==2.0.10
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
PyJWT==2.10.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytorch-lightning==2.5.2
pytorch_toolbelt==0.8.0
pytz==2025.2
pywin32==311
PyYAML==6.0.2
regex==2025.7.34
requests==2.32.4
safetensors==0.5.3
scikit-image==0.25.2
scikit-learn==1.2.2
scipy==1.10.1
segmentation-models-pytorch==0.3.2
sentry-sdk==2.34.1
simsimd==6.5.0
six==1.17.0
smmap==5.0.2
sqlparse==0.5.3
stringzilla==3.12.5
sympy==1.13.3
tabulate==0.9.0
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorboardX==2.6.4
termcolor==3.1.0
threadpoolctl==3.6.0
tifffile==2025.6.11
timm==0.6.13
tokenizers==0.19.1
# torch==2.1.0+cu121
# torchaudio==2.1.0+cu1216
# torchgeometry==0.1.2
# torchmetrics==1.8.0
# torchvision==0.16.0+cu121
tqdm==4.67.1
transformers==4.41.2
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
wandb==0.21.0
Werkzeug==3.1.3
yacs==0.1.8
yarl==1.20.1
zipp==3.23.0
